+++ irbuilder.cabal
@@ -1,4 +1,4 @@
--- This file has been generated from package.yaml by hpack version 0.19.3.
+-- This file has been generated from package.yaml by hpack version 0.17.1.
 --
 -- see: https://github.com/sol/hpack
 
@@ -6,23 +6,30 @@ name:           irbuilder
 version:        0.1.0.0
 author:         Stephen Diehl, Olle Fredriksson
 license:        MIT
+license-file:   LICENSE
 build-type:     Simple
 cabal-version:  >= 1.10
 
 library
+  hs-source-dirs:
+      src
   ghc-options: -fwarn-incomplete-patterns
   build-depends:
       base ==4.9.*
-    , bytestring
-    , containers
-    , llvm-hs-pretty
-    , llvm-hs-pure
-    , mtl
-    , unordered-containers
-    , text
     , transformers
+    , mtl
+    , bytestring
+    , text
+    , containers
+    , unordered-containers
+    , llvm-hs-pure
+    , llvm-hs-pretty
   exposed-modules:
       IRBuilder
+      IRBuilder.Monad
+      IRBuilder.Instruction
+      ModuleBuilder
+      Util.SnocList
   other-modules:
       Paths_irbuilder
   default-language: Haskell2010
